+++ ExpenseTracking.Api/Impl/Cqrs/Address.cs
@@ -7,7 +7,7 @@ using MediatR;
 namespace ExpenseTracking.Api.Impl.Cqrs;
 
 public record GetAllAdressesQuery : IRequest<ApiResponse<List<AddressResponse>>>;
-public record GetAdressesByIdQuery(int Id) : IRequest<ApiResponse<AddressResponse>>;
+public record GetAdressesByIdQuery(long Id) : IRequest<ApiResponse<AddressResponse>>;
 public record GetAdressesByParametersQuery(string? City, string? ZipCode, bool? IsDefault) : IRequest<ApiResponse<List<AddressResponse>>>;
 
 public record CreateAddressForEmployeeCommand(long EmployeeId, AddressRequest Address)
@@ -21,5 +21,5 @@ public record CreateAddressForDepartmentCommand(long DepartmentId, AddressReques
   public void ApplyOwner(Address e) => e.DepartmentId = DepartmentId;
 }
 
-public record UpdateAddressCommand(int Id, AddressRequest Address) : IRequest<ApiResponse>;
-public record DeleteAddressCommand(int Id) : IRequest<ApiResponse>;
+public record UpdateAddressCommand(long Id, AddressRequest Address) : IRequest<ApiResponse>;
+public record DeleteAddressCommand(long Id) : IRequest<ApiResponse>;
